PROD-33717: Add a new generic type for adding data as an entity reference
In some cases we'll display an entity in an object e.g. data->group or data->topic.
In other cases where there is a reference needed, we store the type in the generic target topic: data->target.
This new type holds the ID, type (e.g. post, or topic), and the canonical href.

We don't want to show node as a type, so we use the bundle in that case. Same for group entities, even though there might not yet be a direct comment field for it.