PROD-33731: Implement the payload design for post creation
We added a new PostContentVisibility type similar to the ContentVisibility type defined in the social_eda module. The reason for doing this is that the visibility for posts is defined in a different way, which makes the ContentVisibility class much more complex. I figured it'd be better to separate it entirely.
Both the PostContentVisibility and Stream types have a `jsonSerialize()` function that ensures during serialization (before we sent the event out to Kafka), we remove unneeded attributes (like user if the stream was group).