PROD-33612: Fix Behat step for user segment creation
The scenario 'Create a simple user segment' has been
temporarily disabled due to the 'I create a user segment
using its creation page' step failing.

This failure occurred after the validation for "Rules" field
was added. This commit fixes the way the value of the
hidden rules field is inserted during the creation of the
user segment entity, as well as applying the expected
data (valid JSON string) to the rules field.