Fix dependencies for modules after moving them on disk
Modules are being installed based on their dependency order and if
there's no dependencies then they'll be installed based on the position
on the file system. This does prioritise modules in the install profile
over modules outside of it.

Since we've moved `social_course` around we now re-order some of the
installation order. This caused missing dependency errors for config.

This commit ensures that `social_group` (the system) exists before
`social_course` (the feature). It also makes explicit that
`social_group` requires `flag` (for the unfollow).

Finally we have to create a dependency in the wrong direction
`social_group_request` (system) onto `social_group_flexible_group`
(feature) because this is now treated as a shared way to add request to
join functionality to group-based features, but was previously an
optional feature for flexible group.