Make sure we filter on Individual scoped gropu roles, because all other roles would be covered by using the global role instead.
We've made sure we can filter on scope.
By default we allow all scopes, but that means for Organizations / Airlines it provides all of the insider and outsider roles.
Admin, anonymous etc. That doesn't make sense for this. Perhaps not in the future either, in that case we can always filter on scope.
But this seems like low impact high potential value to filter on scope.