PROD-33775: Add a tiny state management service for group memberships
Group membership requests are group content entities, just like memberships are.
We have a challenge though. When a request is accepted it also creates the actual group membership. Which in turn fires of the "direct join" event.
That's not something we want.

This service is very simple, it record whether a request to join is accepted in the key value expire store. We only need it for <10 seconds as we'll use this in the "direct join" event trigger to check whether it should be fired or not.