PROD-33793: Introduce a payload event for page views for authenticated users
This follows the payload design from PROD-33541.

The aim of this initial page view tracking is to track JUST authenticated users. We do not care about anonymous traffic as it cannot be reliably measured, as we aren't using any JS based tracking.
We aim to get a simple payload that gives us the information about the current URL, who is performing the page view action, and where possible an entity reference.

The latter is an array, but for now capture only a few specific entities (group, node, user, comment, etc.). We don't aim to be complete yet as we expect to add more and more in the future.
With no "entity" know, e.g. now with overview pages, it's NULL.