PROD-33753: Require top-level Rule/RuleGroup in back-end
We previously changed the front-end to send a RuleGroup rather than an
object with a `rules` property to the back-end.

This commit now also makes the change on the back-end. This can simplify
a bunch of code because we can assume we have a uniform tree of `Rule`
or `RuleGroup` options and we do not need to create a special case for
the top-level object.