PROD-34053: Ignore the user and group entities updates for activity entity cache invalidation
Saving a user with many activities associated with them (we've seen cases of >12K) adds a cache invalidation for all of them.
Similarly this happens also for groups. In the end causing 504 errors due to page load times of over 60 seconds (at the EC).

It's very likely that this is entire entity update cache invalidation part is not needed anymore. This requires more investigation and testing.
For now this hotfix addresses the performance issue, rather than fixing the entire thing.