PROD-33753: Convert UserSegment rules field to strict type
We created very nice DataTransfer objects to represent our rules
structure. However, in a lot of places we were still working directly
with JSON, instead of ensuring that we transform the data at the edge
between our front-end and back-end and again between our back-end and
database.

Rather than storing JSON from the front-end directly and validating it
once we start using it, we require that the UserSegment is created with
validated data already so that we know that anything that's loaded will
work.

The rules field is made required since it does not make sense to have
a segment without rules (and we had already implemented validation
for this in the front-end). This simplifies the serialization logic
within the entity.

The storage format has been switched from JSON to a serialized PHP
object. We had originally chosen JSON with the idea that we could
then query against the structure. This breaks down for two reasons:

1. Drupal does not yet support JSON as a data type itself so the
   query API is not ready for this.
2. The rule configuration is a tree structure and it's unclear
   whether the JSON field type would make it easy to query against
   such data anyway.

Ensuring that we have structured data in the field required jumping
around some hoops within the Field API (and FieldItem magic) and in
that scenario it was far easier to use Drupal's serialization, then
try to also take this over: we can send structured data to Drupal
and it'll serialize it, to then give us back the serialized string
that we can hydrate with known classes; however, to work with JSON
we'd have to make the field item convert between structured data
and JSON on the fly and do double bookkeeping. That's likely
possible, but much more work than we currently had time for.

The validation of User Segment plugin IDs is removed as this is not a
user entity/storage concern but is a concern for the consumers. For
example a plugin might be removed while data is stored in the database
which could break changing the label. This makes the behaviour
consistent with other plugin usages in Drupal.