Remove .github folder from distribution
All relevant workflows and rules have been migrated to our main
monorepo. Changes are sometimes made in the `.github` folder (e.g.
because of outdated workflows). However, for safety reasons, our subtree
split automation is not allowed to update the .github folder since this
could pose a security risk. We could work around that with higher
permissions, but we don't want to give those to our workflow.

The easiest solution is to clean up the `.github` folder altogether so
we avoid the associated maintenance and conflicts.