Gaia AI spam prevention
This combines the power of Gaia's relevancy score, as a threshold for spam prevention measures.
Where, based on the relevancy score, actions are taken. For now those actions are, blocking the author of the entity, unpublishing the entity, and extract spam words and add those to the block list at social_spam_prevention if that module is enabled. This is all pluggable.

This allows a SM to configure the thresholds per pluging through a menu link in the dashboard.
For example when the score of a given entity is below 40, the entity is unpublished.
And/or if the score is below 20, the user gets blocked etc.

This is a first rough PoC to showcase the opportunities we have and add value to Gaia in the light of the spam issues we have currently.