Configure known entity types in PHPStan
PHPStan is really smart if we tell it about the entity types that exist
and their corresponding storage classes. This allows us to remove a lot
of type checking code in our codebase. For example because it knows that
any entity loaded into an array are of a certain type.