Clear Redis cache on database load in Behat tests
We've seen that in some test cases the Redis cache may still contain
data from previous tests and this causes flakiness. There's an issue for
this open but it's waiting for the Redis maintainer.

This commit includes the code from the proposed PR in a context to help
clear the Redis caches when an install database is loaded. This aims to
reduce the flakiness of tests.