Ensure `config_modify` module dependency is properly declared
We're using the `config_modify` module in our project to make sure that
we can easily do partial config updates. Since we currently install Open
Social with a small set of modules first and only then start enabling a
lot of other modules, config_modify is usually available when modules
using it are enabled.

However, this is not guaranteed and we want to let Drupal enable more
modules during an initial site-install to standardise our set-up
process. In those scenarios, if the dependency on config_modify is not
declared, then it may be enabled after a module that provides config
modifications. In those scenarios the config_modify module assumes that
they're already applied and will skip them. This can lead to
inconsistent behaviour.

To make sure that all config modifications are applied as expected,
modules that provide `config/modify` files should declare a dependency
on `config_modify`. This ensures that Drupal enabled the config_modify
module early in the process of installation and that config gets applied
as expected. This is similar to how `config/install` or
`config/optional` expects Drupal's built in config installer to always
be available.

To make sure this doesn't break again in the future we add a simple
validation command to conductor and wire this up in our CI.