Fix Social Book permission assignment during install
The book node type is optional config in the `book` module. The Drupal
installer optimizes config install by first installing all required
configuration and only later installing optional config. This means
that when `social_book_install` runs during installation the book
content type doesn't actually exist yet.

This means that when the permissions are assigned in the install hook,
because the book content type does not yet exist, the permissions are
invalid. This causes Drupal to remove these permissions from the roles.

By moving the permissions into config/modify we make things run after
optional config is installed. Additionally it requires us to set a
dependency on the book content type to exist, so that even if it would
be made available even later then at that point in time the permissions
will be assigned.