PROD-34566: Add Event Enrollment Request BackfillHandler plugins for EDA backfill operations
The base class is already named for both requests and invites, as this supports also the event enrollment invites (uses same condition check, different value).

We split it up into two different plugins, each focusing on a state. Either "pending" (create), or "approved" (accept).
We do not support  the request declined action, as the status is the same for both event request declined and invited declined. Making it impossible to determine which of the two events to fire.