PROD-34609: Introduce a way to set the actor for an event during backfill operations
We want to avoid changing the business logic of the current event generation as much as possible. But leaving the actor information empty, while we do have reliable information is a missed chance.

This change introduces a way to override the "currentUser" information set during the event creation. With the use of an interface we can have simple type checking.
We can let each individual backfill handler plugin resolve the actor user itself, while using common methods to override the change.